app/vallidator-bff/python/requirements.txt
flask
pandas
openpyxl
xlrd
google-cloud-storage
psycopg2-binary
